[Themes] Added ProvisionedPathResolver + AssetPathResolutionPass
Goal is to browse all theme folders and pre-resolve paths for found assets.
This will avoid multiple I/O access at runtime in production.

A setting has been added in order to enable/disable this feature.
Default behavior is to enable it only when debug is off (i.e. "production" environment).

Also started to change services names, in order to prepare extraction to a specific bundle.